Fix some inconsistency with nsvr_region, remove dead code

+++ src/Driver/SharedCommunication/SharedTypes.h
@@ -86,7 +86,7 @@ namespace NullSpace {
 			using RegionType = uint32_t;
 			RegionType Region;
 			uint32_t Type;
-			uint64_t Id;
+			uint32_t Id;
 			Data Value;
